python/2023/day17/test_day17.py
import pytest
import day17 as aoc


@pytest.fixture
def example1():
    puzzle_input = """2413432311323
3215453535623
3255245654254
3446585845452
4546657867536
1438598798454
4457876987766
3637877979653
4654967986887
4564679986453
1224686865563
2546548887735
4322674655533"""
    return aoc.parse(puzzle_input)


@pytest.fixture
def example2():
    puzzle_input = """   """
    return aoc.parse(puzzle_input)


@pytest.mark.skip(reason="Not implemented")
def test_parse_example1(example1):
    """Test that input is parsed properly."""
    assert example1 == ...


def test_part1_example1(example1):
    """Test part 1 on example input."""
    assert aoc.part1(example1) == 102


def test_part2_example1(example1):
    """Test part 2 on example input."""
    assert aoc.part2(example1) == 94


@pytest.mark.skip(reason="Not implemented")
def test_part2_example2(example2):
    """Test part 2 on example input."""
    assert aoc.part2(example2) == ...
